feat : question vote up down기능 ver 2.1

+++ BE/project/src/main/java/com/example/project/question/dto/QuestionDto.java
@@ -107,7 +107,7 @@ public class QuestionDto {
     @NoArgsConstructor
     public static class RecommendResponseDto {
 
-        private int voteCount;
+        private long voteCount;
 
     }
